Enhance implementation of fragment to accelerate the compilation
