+++ modules/graph/loader/basic_ev_fragment_loader_int32.cc
@@ -13,13 +13,17 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#ifndef MODULES_GRAPH_FRAGMENT_ARROW_FRAGMENT_MODIFIER_H_
-#define MODULES_GRAPH_FRAGMENT_ARROW_FRAGMENT_MODIFIER_H_
+#include "graph/loader/basic_ev_fragment_loader_impl.h"
+#include "graph/utils/partitioner.h"
 
 namespace vineyard {
 
-// keep for backwards compatibility.
+template class BasicEVFragmentLoader<
+    int32_t, uint64_t, HashPartitioner<int32_t>,
+    ArrowVertexMap<typename InternalType<int32_t>::type, uint64_t>>;
+
+template class BasicEVFragmentLoader<
+    int32_t, uint64_t, HashPartitioner<int32_t>,
+    ArrowLocalVertexMap<typename InternalType<int32_t>::type, uint64_t>>;
 
 }  // namespace vineyard
-
-#endif  // MODULES_GRAPH_FRAGMENT_ARROW_FRAGMENT_MODIFIER_H_
