+++ modules/graph/loader/basic_ev_fragment_loader_string.cc
@@ -13,13 +13,17 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#ifndef MODULES_GRAPH_FRAGMENT_ARROW_FRAGMENT_MODIFIER_H_
-#define MODULES_GRAPH_FRAGMENT_ARROW_FRAGMENT_MODIFIER_H_
+#include "graph/loader/basic_ev_fragment_loader_impl.h"
+#include "graph/utils/partitioner.h"
 
 namespace vineyard {
 
-// keep for backwards compatibility.
+template class BasicEVFragmentLoader<
+    std::string, uint64_t, HashPartitioner<std::string>,
+    ArrowVertexMap<typename InternalType<std::string>::type, uint64_t>>;
+
+template class BasicEVFragmentLoader<
+    std::string, uint64_t, HashPartitioner<std::string>,
+    ArrowLocalVertexMap<typename InternalType<std::string>::type, uint64_t>>;
 
 }  // namespace vineyard
-
-#endif  // MODULES_GRAPH_FRAGMENT_ARROW_FRAGMENT_MODIFIER_H_
